+++ src/main/kotlin/com/lob/postcard/Application.kt
@@ -3,11 +3,9 @@ package com.lob.postcard
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.runApplication
 
-@Suppress("SpreadOperator")
 fun main(args: Array<String>) {
     runApplication<Application>(*args)
 }
 
 @SpringBootApplication(scanBasePackages = ["com.lob.postcard"])
-@Suppress("SpreadOperator")
 open class Application
