+++ src/main/kotlin/com/lob/postcard/api/PostCardController.kt
@@ -25,6 +25,7 @@ class PostCardController(
 ) {
     @GetMapping(
         value = ["/postcard/addresses"],
+        consumes = ["application/json"],
         produces = ["application/json"]
     )
     fun getAddresses(
