+++ src/main/resources/addresses.json
@@ -1,0 +1,53 @@
+[
+  {
+    "line1": "Massachusetts Hall",
+    "city": "Cambridge",
+    "state": "MA",
+    "zip": "02138"
+  },
+  {
+    "line1": "3400 N. Charles St.",
+    "city": "Baltimore",
+    "state": "MD",
+    "zip": "21218"
+  },
+  {
+    "line1": "Roosevelt Way NE",
+    "city": "Seattle",
+    "state": "WA",
+    "zip": "98115"
+  },
+  {
+    "line1": "1600 Holloway Ave",
+    "city": "San Francisco",
+    "state": "CA",
+    "zip": "94132"
+  },
+  {
+    "line1": "1600 Holloway Ave",
+    "line2": "Suite 10",
+    "city": "San Francisco",
+    "state": "CA",
+    "zip": "94132"
+  },
+  {
+    "line1": "1600 Holloway Ave",
+    "line2": "Suite 20",
+    "city": "San Francisco",
+    "state": "CA",
+    "zip": "94132"
+  },
+  {
+    "line1": "500 S State St",
+    "city": "Ann Arbor",
+    "state": "MI",
+    "zip": "48109"
+  },
+  {
+    "line1": "185 Berry St",
+    "line2": "Suite 6100",
+    "city": "San Francisco",
+    "state": "CA",
+    "zip": "94107"
+  }
+]
